refactor(layerChooser): add capabilities module directly to the layerChooser store
